requirements.txt
sqlalchemy==1.4.32
psycopg2-binary
python-dotenv
